force beta_output to be torch.float32

+++ vllm/model_executor/models/olmo3_5_hybrid.py
@@ -1152,7 +1152,7 @@ def fused_olmo3_5_hybrid_gdn_gating(
     seq_len = 1
     grid = (batch, seq_len, triton.cdiv(num_heads, 8))
     g = torch.empty(1, batch, num_heads, dtype=torch.float32, device=a.device)
-    beta_output = torch.empty(1, batch, num_heads, dtype=b.dtype, device=b.device)
+    beta_output = torch.empty(1, batch, num_heads, dtype=torch.float32, device=b.device)
     fused_olmo3_5_hybrid_gdn_gating_kernel[grid](
         g,
         beta_output,
